// src/api/configuration/meterTypeParameters/types.d.ts
type MeterTypeParam = {
    CREATE_TIME: string;
    DB_DOT: string;
    DEFAULT_VALUE: string;
    EDIT_TYPE: string;
    ENUM_LIST: string;
    ENUM_LIST_USE: string;
    MAX_VALUE: string;
    METER_MODEL_NO: string;
    MIN_VALUE: string;
    MODIFY_TIME: string;
    PARAM_XUHAO: string;
    PROTOCOL_NAME: string;
    PROTOCOL_NAME_USE: string;
    PROTOCOL_NO: string;
    IS_HIDDEN: string;
    DATA_TYPE: string;
    XU_HAO: string;
    SN: string;
    PROTOCAL_NAME: string;
    CARD_TYPENO: string;
}

export {
    MeterTypeParam,
};